gradio-tutorial/requirements.txt
gradio
langchain-community 
langchain 
langchain-google-genai 
python-dotenv